refactor!: move configuration into a unified module
This is an instance where we exposed different interfaces to the
different targets, so I just did what I could to unify them.

# Conflicts:
#	crypto-ffi/src/generic/mod.rs
