+++ crypto-ffi/src/wasm/mod.rs
@@ -55,109 +55,6 @@ impl From<core_crypto::e2e_identity::E2eiDumpedPkiEnv> for E2eiDumpedPkiEnv {
     }
 }
 
-#[wasm_bindgen]
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-/// Configuration object for new conversations
-/// see [core_crypto::prelude::MlsConversationConfiguration]
-pub struct ConversationConfiguration {
-    #[wasm_bindgen(readonly)]
-    /// Conversation ciphersuite
-    pub ciphersuite: Option<Ciphersuite>,
-    external_senders: Vec<Vec<u8>>,
-    #[wasm_bindgen(readonly)]
-    /// Additional configuration
-    pub custom: CustomConfiguration,
-}
-
-#[wasm_bindgen]
-impl ConversationConfiguration {
-    #[wasm_bindgen(constructor)]
-    pub fn new(
-        ciphersuite: Option<Ciphersuite>,
-        external_senders: Option<Vec<Uint8Array>>,
-        key_rotation_span: Option<u32>,
-        wire_policy: Option<WirePolicy>,
-    ) -> WasmCryptoResult<ConversationConfiguration> {
-        let external_senders = external_senders
-            .map(|exs| exs.iter().cloned().map(|jsv| jsv.to_vec()).collect())
-            .unwrap_or_default();
-        Ok(Self {
-            ciphersuite,
-            external_senders,
-            custom: CustomConfiguration::new(key_rotation_span, wire_policy),
-        })
-    }
-
-    /// List of client IDs that are allowed to be external senders
-    #[wasm_bindgen(getter, js_name = externalSenders)]
-    pub fn external_senders(&self) -> js_sys::Array {
-        self.external_senders
-            .iter()
-            .cloned()
-            .map(JsValue::from)
-            .collect::<js_sys::Array>()
-    }
-}
-
-#[wasm_bindgen]
-#[derive(Debug, Copy, Clone, serde::Serialize, serde::Deserialize)]
-/// see [core_crypto::prelude::MlsCustomConfiguration]
-pub struct CustomConfiguration {
-    ///  Duration in seconds after which we will automatically force a self-update commit
-    ///  Note: This isn't currently implemented
-    #[wasm_bindgen(js_name = keyRotationSpan)]
-    pub key_rotation_span: Option<u32>,
-    /// Defines if handshake messages are encrypted or not
-    /// Note: encrypted handshake messages are not supported by wire-server
-    #[wasm_bindgen(js_name = wirePolicy)]
-    pub wire_policy: Option<WirePolicy>,
-}
-
-#[wasm_bindgen]
-impl CustomConfiguration {
-    #[wasm_bindgen(constructor)]
-    pub fn new(key_rotation_span: Option<u32>, wire_policy: Option<WirePolicy>) -> Self {
-        Self {
-            key_rotation_span,
-            wire_policy,
-        }
-    }
-}
-
-impl From<CustomConfiguration> for MlsCustomConfiguration {
-    fn from(cfg: CustomConfiguration) -> Self {
-        let key_rotation_span = cfg
-            .key_rotation_span
-            .map(|span| std::time::Duration::from_secs(span as u64));
-        let wire_policy = cfg.wire_policy.map(WirePolicy::into).unwrap_or_default();
-        Self {
-            key_rotation_span,
-            wire_policy,
-            ..Default::default()
-        }
-    }
-}
-
-#[wasm_bindgen]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-#[repr(u16)]
-/// see [core_crypto::prelude::MlsWirePolicy]
-pub enum WirePolicy {
-    /// Handshake messages are never encrypted
-    Plaintext = 0x0001,
-    /// Handshake messages are always encrypted
-    Ciphertext = 0x0002,
-}
-
-impl From<WirePolicy> for MlsWirePolicy {
-    fn from(policy: WirePolicy) -> Self {
-        match policy {
-            WirePolicy::Plaintext => Self::Plaintext,
-            WirePolicy::Ciphertext => Self::Ciphertext,
-        }
-    }
-}
-
 static INIT_LOGGER: Once = Once::new();
 static LOGGER: LazyLock<ReloadLog<CoreCryptoWasmLogger>> = LazyLock::new(|| {
     ReloadLog::new(CoreCryptoWasmLogger {
